+++ src/QuickMarcEditor/QuickMarcEditorContainer.js
@@ -85,7 +85,6 @@ const QuickMarcEditorContainer = ({
   const [marcRecord, setMarcRecord] = useState();
   const [locations, setLocations] = useState();
   const [isLoading, setIsLoading] = useState(true);
-  const [linksCount, setLinksCount] = useState(0);
   const [fixedFieldSpec, setFixedFieldSpec] = useState();
 
   const searchParams = new URLSearchParams(location.search);
@@ -96,7 +95,7 @@ const QuickMarcEditorContainer = ({
     && action !== QUICK_MARC_ACTIONS.CREATE;
 
   const showCallout = useShowCallout();
-  const { fetchLinksCount } = useAuthorityLinksCount();
+  const { linksCount } = useAuthorityLinksCount({ id: marcType === MARC_TYPES.AUTHORITY && externalId });
 
   const closeEditor = useCallback((id) => {
     if (marcType === MARC_TYPES.HOLDINGS && action !== QUICK_MARC_ACTIONS.CREATE) {
@@ -140,13 +139,12 @@ const QuickMarcEditorContainer = ({
         ...headers,
       });
 
-    const locationsPromise = mutator.locations.GET(headers);
+    const locationsPromise = marcType === MARC_TYPES.HOLDINGS
+      ? mutator.locations.GET(headers)
+      : Promise.resolve();
+
     // must be with the central tenant id when user derives shared record
     const linkingRulesPromise = mutator.linkingRules.GET(headers);
-
-    const linksCountPromise = marcType === MARC_TYPES.AUTHORITY
-      ? fetchLinksCount([externalId])
-      : Promise.resolve();
 
     const fixedFieldSpecPromise = mutator.fixedFieldSpec.GET({
       path: `${MARC_SPEC_API}/${marcType}/008`,
@@ -157,7 +155,6 @@ const QuickMarcEditorContainer = ({
       instancePromise,
       marcRecordPromise,
       locationsPromise,
-      linksCountPromise,
       linkingRulesPromise,
       fixedFieldSpecPromise,
     ])
@@ -165,14 +162,9 @@ const QuickMarcEditorContainer = ({
         instanceResponse,
         marcRecordResponse,
         locationsResponse,
-        linksCountResponse,
         linkingRulesResponse,
         fixedFieldSpecResponse,
       ]) => {
-        if (marcType === MARC_TYPES.AUTHORITY) {
-          setLinksCount(linksCountResponse.links[0].totalLinks);
-        }
-
         if (action !== QUICK_MARC_ACTIONS.CREATE) {
           searchParams.set('relatedRecordVersion', instanceResponse._version);
 
@@ -184,7 +176,7 @@ const QuickMarcEditorContainer = ({
         let dehydratedMarcRecord;
 
         if (action === QUICK_MARC_ACTIONS.CREATE) {
-          dehydratedMarcRecord = createRecordDefaults[marcType](instanceResponse);
+          dehydratedMarcRecord = createRecordDefaults[marcType](instanceResponse, fixedFieldSpecResponse);
         } else {
           dehydratedMarcRecord = dehydrateMarcRecordResponse(marcRecordResponse, marcType, fixedFieldSpecResponse);
         }
@@ -208,7 +200,6 @@ const QuickMarcEditorContainer = ({
     externalId,
     history,
     marcType,
-    fetchLinksCount,
     centralTenantId,
     token,
     locale,