+++ src/QuickMarcEditor/QuickMarcEditorRows/QuickMarcEditorRows.test.js
@@ -258,7 +258,7 @@ describe('Given QuickMarcEditorRows', () => {
     describe('when the subfield is focused', () => {
       it('should have a cursor at the end of the field value', () => {
         const { getByTestId } = renderQuickMarcEditorRows();
-        const subfield = getByTestId('content-field-0');
+        const subfield = getByTestId('content-field-4');
         const valueLength = subfield.value.length;
         const spySetSelectionRange = jest.spyOn(subfield, 'setSelectionRange');