+++ src/QuickMarcEditor/utils.js
@@ -33,6 +33,8 @@ import {
   CREATE_AUTHORITY_RECORD_DEFAULT_FIELD_TAGS,
   UNCONTROLLED_ALPHA,
   UNCONTROLLED_NUMBER,
+  TAG_LENGTH,
+  AUTHORITY_FIXED_FIELD_DEFAULT_TYPE,
 } from './constants';
 import { RECORD_STATUS_NEW } from './QuickMarcRecordInfo/constants';
 import { SUBFIELD_TYPES } from './QuickMarcEditorRows/BytesField';
@@ -55,6 +57,7 @@ export const isLastRecord = recordRow => {
   );
 };
 
+export const isControlNumberRow = recordRow => recordRow.tag === '001';
 export const isFixedFieldRow = recordRow => recordRow.tag === '008';
 export const isMaterialCharsRecord = recordRow => recordRow.tag === '006';
 export const isPhysDescriptionRecord = recordRow => recordRow.tag === '007';
@@ -63,7 +66,8 @@ export const isContentRow = (recordRow, marcType) => {
   return !(isLocationRow(recordRow, marcType)
     || isFixedFieldRow(recordRow)
     || isMaterialCharsRecord(recordRow)
-    || isPhysDescriptionRecord(recordRow));
+    || isPhysDescriptionRecord(recordRow)
+    || isControlNumberRow(recordRow));
 };
 
 export const getContentSubfieldValue = (content = '') => {
@@ -246,10 +250,10 @@ const getCreateMarcRecordDefaultFields = (contentMap, indicatorMap, defaultTags)
   });
 };
 
-const getCreateBibMarcRecordDefaultFields = (instanceRecord) => {
+const getCreateBibMarcRecordDefaultFields = (instanceRecord, fixedFieldSpec) => {
   const contentMap = {
     '001': instanceRecord.hrid,
-    '008': fillEmptyFixedFieldValues(MARC_TYPES.BIB, BIB_FIXED_FIELD_DEFAULT_TYPE, BIB_FIXED_FIELD_DEFAULT_BLVL),
+    '008': fillEmptyFixedFieldValues(MARC_TYPES.BIB, fixedFieldSpec, BIB_FIXED_FIELD_DEFAULT_TYPE, BIB_FIXED_FIELD_DEFAULT_BLVL),
     '245': '$a ',
     '999': '',
   };
@@ -276,16 +280,14 @@ const getCreateHoldingsMarcRecordDefaultFields = (instanceRecord) => {
   return getCreateMarcRecordDefaultFields(contentMap, indicatorMap, CREATE_HOLDINGS_RECORD_DEFAULT_FIELD_TAGS);
 };
 
-const getCreateAuthorityMarcRecordDefaultFields = (instanceRecord) => {
+const getCreateAuthorityMarcRecordDefaultFields = (instanceRecord, fixedFieldSpec) => {
   const contentMap = {
     '001': instanceRecord.hrid,
-    '008': fillEmptyFixedFieldValues(MARC_TYPES.AUTHORITY),
-    '100': '$a ',
+    '008': fillEmptyFixedFieldValues(MARC_TYPES.AUTHORITY, fixedFieldSpec, AUTHORITY_FIXED_FIELD_DEFAULT_TYPE),
     '999': '',
   };
 
   const indicatorMap = {
-    '100': ['\\', '\\'],
     '999': ['f', 'f'],
   };
 
@@ -311,7 +313,7 @@ export const getCreateHoldingsMarcRecordResponse = (instanceResponse) => {
   };
 };
 
-export const getCreateBibMarcRecordResponse = (instanceResponse) => {
+export const getCreateBibMarcRecordResponse = (instanceResponse, fixedFieldSpec) => {
   const instanceId = '00000000-0000-0000-0000-000000000000'; // For create we need to send any UUID
 
   return {
@@ -324,13 +326,13 @@ export const getCreateBibMarcRecordResponse = (instanceResponse) => {
         content: CREATE_BIB_RECORD_DEFAULT_LEADER_VALUE,
         id: LEADER_TAG,
       },
-      ...getCreateBibMarcRecordDefaultFields(instanceResponse),
+      ...getCreateBibMarcRecordDefaultFields(instanceResponse, fixedFieldSpec),
     ],
     parsedRecordDtoId: instanceId,
   };
 };
 
-export const getCreateAuthorityMarcRecordResponse = (instanceResponse) => {
+export const getCreateAuthorityMarcRecordResponse = (instanceResponse, fixedFieldSpec) => {
   const instanceId = '00000000-0000-0000-0000-000000000000'; // For create we need to send any UUID
 
   return {
@@ -343,7 +345,7 @@ export const getCreateAuthorityMarcRecordResponse = (instanceResponse) => {
         content: CREATE_AUTHORITY_RECORD_DEFAULT_LEADER_VALUE,
         id: LEADER_TAG,
       },
-      ...getCreateAuthorityMarcRecordDefaultFields(instanceResponse),
+      ...getCreateAuthorityMarcRecordDefaultFields(instanceResponse, fixedFieldSpec),
     ],
     parsedRecordDtoId: instanceId,
   };
@@ -474,6 +476,14 @@ export const getLocationValue = (value) => {
   const matches = value?.match(/\$b\s+([^$\s]+\/?)+/) || [];
 
   return matches[0] || '';
+};
+
+export const checkIsEmptyContent = (field) => {
+  if (typeof field.content === 'string') {
+    return compact(field.content.split(' ')).every(content => /^\$[a-z0-9]?$/.test(content));
+  }
+
+  return false;
 };
 
 export const checkIsInitialRecord = (field) => {
@@ -680,14 +690,6 @@ export const removeFieldsForDerive = (formValues) => {
   };
 };
 
-const checkIsEmptyContent = (field) => {
-  if (typeof field.content === 'string') {
-    return compact(field.content.split(' ')).every(content => /^\$[a-z0-9]?$/.test(content));
-  }
-
-  return false;
-};
-
 export const autopopulateIndicators = (formValues) => {
   const { records } = formValues;
 
@@ -750,7 +752,7 @@ export const autopopulatePhysDescriptionField = (formValues) => {
   };
 };
 
-export const autopopulateFixedField = (formValues, marcType, marcSpec) => {
+export const autopopulateFixedField = (formValues, marcType, fixedFieldSpec) => {
   const { records } = formValues;
 
   const leader = records.find(field => field.tag === LEADER_TAG);
@@ -766,7 +768,7 @@ export const autopopulateFixedField = (formValues, marcType, marcSpec) => {
 
       return {
         ...field,
-        content: fillEmptyFixedFieldValues(marcType, marcSpec, type, blvl, field),
+        content: fillEmptyFixedFieldValues(marcType, fixedFieldSpec, type, blvl, field),
       };
     }),
   };
@@ -1072,10 +1074,10 @@ const addLeaderFieldAndIdToRecords = (marcRecordResponse) => ({
   ],
 });
 
-export const dehydrateMarcRecordResponse = (marcRecordResponse, marcType, marcSpec) => (
+export const dehydrateMarcRecordResponse = (marcRecordResponse, marcType, fixedFieldSpec) => (
   flow(
     addLeaderFieldAndIdToRecords,
-    marcRecord => autopopulateFixedField(marcRecord, marcType, marcSpec),
+    marcRecord => autopopulateFixedField(marcRecord, marcType, fixedFieldSpec),
     autopopulatePhysDescriptionField,
     autopopulateMaterialCharsField,
   )(marcRecordResponse)