+++ src/hooks/useValidation/validators.test.js
@@ -1,0 +1,11 @@
+import * as validators from './validators';
+
+const locations = [{
+  code: 'VA/LI/D',
+}, {
+  code: 'LO/CA/TI/ON',
+}];
+
+describe('validators', () => {
+  
+});